+++ src/backend/community-service/src/main/java/harmony/communityservice/guild/category/dto/SearchCategoryResponse.java
@@ -3,5 +3,5 @@ package harmony.communityservice.guild.category.dto;
 import lombok.Builder;
 
 @Builder(toBuilder = true)
-public record SearchCategoryResponse(int categoryId, Long guildId, String categoryName) {
+public record SearchCategoryResponse(Long categoryId, Long guildId, String categoryName) {
 }
